LayoutTests/platform/ios-simulator/fast/dom/HTMLImageElement/sizes/image-sizes-2x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
FAIL document.getElementById("simple").clientWidth should be 320. Was 800.
PASS currentSrcFileName("simple") is "image-set-2x.png"
PASS document.getElementById("small_sizes").clientWidth is 300
PASS currentSrcFileName("small_sizes") is "image-set-2x.png"
PASS document.getElementById("sizes_x_descriptors").clientWidth is 400
PASS currentSrcFileName("sizes_x_descriptors") is "image-set-2x.png"
PASS document.getElementById("sizes_src").clientWidth is 400
PASS currentSrcFileName("sizes_src") is "image-set-1x.png"
PASS document.getElementById("default_sizes").clientWidth is window.innerWidth
FAIL currentSrcFileName("default_sizes") should be image-set-4x.png. Was image-set-2x.png.
PASS document.getElementById("invalid_length").clientWidth is window.innerWidth
PASS currentSrcFileName("invalid_length") is "image-set-2x.png"
PASS document.getElementById("src_and_w").clientWidth is within 0.015625 of 700
PASS currentSrcFileName("src_and_w") is "image-set-4x.png"
PASS document.getElementById("x_and_w").clientWidth is within 0.015625 of 700
PASS currentSrcFileName("x_and_w") is "image-set-4x.png"
PASS document.getElementById("x_and_w2").clientWidth is 400
PASS currentSrcFileName("x_and_w2") is "image-set-2x.png"
PASS document.getElementById("x_and_w3").clientWidth is 400
PASS currentSrcFileName("x_and_w3") is "image-set-2x.png"
PASS document.getElementById("half_size").clientWidth is (window.innerWidth/2)
PASS currentSrcFileName("half_size") is "image-set-2x.png"
FAIL document.getElementById("infitisimal_sizes").clientWidth should be 3.2. Was 3.
PASS currentSrcFileName("infitisimal_sizes") is "image-set-2x.png"
              
